tools/pyproject.toml
[tool.poetry]
name = "trainings"
version = "0.1.0"
description = ""
authors = ["bastantoine <bastien.antoine69+github@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-frontmatter = "^1.1.0"
loguru = "^0.7.2"
click = "^8.1.7"
jinja2 = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
